requirements-test.txt
# RadioCalico Testing Dependencies (PostgreSQL)
# Add these to your requirements.txt or install separately

pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0
psycopg2-binary==2.9.9
